Add files via upload
About the Project:
The company, in correspondence to the orders received,
releases a weekly plan for production of different models of air
conditioners on different assembly lines. When we consider one
assembly line, there are various products that have to be produced
on that. All the products have their own cycle time, that is time
elapsed between two subsequent ready products. The problem
arises when two consequent products have a large cycle time
difference. Whenever a product with higher cycle time enters the
first station of an assembly line which was previously manufacturing
a product of lesser cycle time, the remaining products of the
previous model are now compelled to run with the higher cycle time.
This results in a lot of time being wasted as the work on the previous
product gets done faster than the latter. This is called change over
loss. However, currently the company goes about it as per the
weekly plan without any ordering leading to drastic change over
losses and thereby decreasing the productivity. The project assigned
to me was to eliminate the manual effort of organising the order of
production on a line by automatically finding the optimised order
with minimum change over loss, through an application software.

The application with a GUI, which fetches a .csv file and opens it for the user to paste
the schedule which contains the name of the model and quantity.
Once the user does this, he clicks on an option to calculate the
optimum schedule where the algorithm fetches the input from the
file, maps it to the corresponding cycle time which is fetched from
another .csv file containing model name and cycle time, calculates
the optimum order and presents the output in another .csv file which
is also the reference file to check for the previous plan’s last cycle
time. The last option view schedule can be clicked to view the
calculated schedule.

I chose to use Java as the programming language and NetBeans
IDE platform. Hence, the final application would be a .jar executable
application with three support .csv files.

The Algorithm for the first scenario, considering only the cycle time is simple though gives tremendous results. There is a variable defined called the reference cycle time (ctr) which holds as the base value for our arrangement. This value is 0 if the algorithm is running for the first time or there is no under-production model on the line. It assumes the value of the cycle time of the last model in the previous schedule if all the items of the last model had reached the line. The last possibility is when the entire schedule is not produced previously or the schedule is updated mid-way, in such cases, the cycle time of the last produced model and its quantity entered by the user is considered as ctr. Now, the entries in the current schedule are arranged in two possibilities w.r.t. the ctr. In the first possibility, the algorithm finds the entries with larger cycle time values than the ctr and arranges it in ascending order till the highest value, later having a steep fall to the largest value of cycle time smaller than ctr and then arranges the later values in descending order. The second possibility is when the entries first dampen till the smallest value in descending order, then a steep rise to the smallest value greater than the ctr and an ascending order rise till the largest cycle time. This gives a continuous wave like curve which ensures there is no sudden changes in cycle time keeping the changeover loss minimum.
For the second scenario, when we need to consider the priority of the orders too, the entries are fed into the same .csv file in the same priority order. The optimum schedule is created using the previous algorithm. Now we have two different orders satisfying two different concerns of the company, but we need to find an intermediate order which lies on the lines of the matching algorithm. We mandate that, the products declared in the first half +1 must be manufactured in the first half +1 with minimum change over loss. Hence we, run a matching algorithm for the first half +1, with their choices of position being the priority position it is in till plus the number of entries/2. The entries in the first half desire to get their best choice or better if an even early position in the final order. The other half -1 is then arranged using the same optimisation algorithm defined above using the last entry in the previous half+1 as the ctr.